item respawn, prevent permanent power up loss, and automated signal connection
Items can now be set to respawn after being collected if they are temporary, allowing for them to be collected again.

Players no longer lose their power up after the time limit if they collect a temporary variant after already having the permanent one.

Automatically connects the item collection signal to the player because it was annoying to do manually.